refactor(datablock): improve  the number of allocation for decode function datablock

+++ item/base.go
@@ -44,7 +44,7 @@ const (
 type UAP struct {
 	Name      string
 	Category  uint8
-	Version   float64
+	Version   float32
 	DataItems []DataItem
 }
